[#1189] (wip) Refactors view logic into serializer modules
* Moves jsvars assignment logic from the base view
  to dedicated serializer modules.
* Improves code organization and separation of concerns.
* Enhances maintainability and readability.
